refactor(db): extract author photo data layer to db/src/author_photos_data.rs
