+++ db/src/queries.rs
@@ -22,204 +22,10 @@ use omnibus_shared::{
 use crate::books::parse_json_array;
 use crate::helpers::build_fts_match;
 pub use crate::helpers::sanitize_fts_query;
-use crate::metadata_overrides::{load_overrides_bulk, rebuild_fts_for_book};
+use crate::metadata_overrides::load_overrides_bulk;
 
 
 
-// -----------------------------------------------------------------------------
-// F1.11 Author profile photos.
-//
-// `author_photos` holds at most one row per author (PK = author_id). The
-// `source` column distinguishes:
-//   - `'manual'`     — admin upload via `PUT /api/authors/:id/photo`.
-//   - `'openlibrary'`— resolved by the background worker.
-//   - `'letter'`     — negative-cache marker: no usable image. `bytes` and
-//                       `mime` are NULL; `get_author_photo` returns `None`
-//                       so the GET handler 404s and the letter avatar
-//                       stays in place.
-// Admin DELETE clears the row to force re-resolution on next view.
-// -----------------------------------------------------------------------------
-
-/// Source-of-truth marker for a cached author photo row.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum AuthorPhotoSource {
-    Manual,
-    OpenLibrary,
-    Letter,
-}
-
-impl AuthorPhotoSource {
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            AuthorPhotoSource::Manual => "manual",
-            AuthorPhotoSource::OpenLibrary => "openlibrary",
-            AuthorPhotoSource::Letter => "letter",
-        }
-    }
-
-    pub fn parse(s: &str) -> Option<Self> {
-        match s {
-            "manual" => Some(Self::Manual),
-            "openlibrary" => Some(Self::OpenLibrary),
-            "letter" => Some(Self::Letter),
-            _ => None,
-        }
-    }
-}
-
-/// Fetch a cached profile photo for the given author. Returns `None` when no
-/// row exists or the row is a `'letter'` negative-cache marker — both cases
-/// should produce a 404 from the GET handler so the frontend keeps rendering
-/// the letter avatar.
-pub async fn get_author_photo(
-    pool: &SqlitePool,
-    author_id: i64,
-) -> Result<Option<(String, Vec<u8>)>, sqlx::Error> {
-    // Filter `letter` rows in SQL as well as via the schema CHECK constraint
-    // so a malformed row (e.g. left over from a future migration drift)
-    // can never accidentally serve `letter` bytes as a real image. Belt +
-    // suspenders alongside the table-level invariant.
-    let row: Option<(String, Option<String>, Option<Vec<u8>>)> = sqlx::query_as(
-        "SELECT source, mime, bytes
-           FROM author_photos
-          WHERE author_id = ? AND source <> 'letter'",
-    )
-    .bind(author_id)
-    .fetch_optional(pool)
-    .await?;
-    match row {
-        Some((_, Some(mime), Some(bytes))) if !bytes.is_empty() => Ok(Some((mime, bytes))),
-        _ => Ok(None),
-    }
-}
-
-/// Look up just the cascade-state metadata (source + fetched_at) for an
-/// author. Used by the resolver to decide whether to skip resolution — a
-/// `'letter'` row prevents re-querying Open Library until an admin clears it
-/// via `delete_author_photo`.
-pub async fn author_photo_status(
-    pool: &SqlitePool,
-    author_id: i64,
-) -> Result<Option<(AuthorPhotoSource, String)>, sqlx::Error> {
-    let row: Option<(String, String)> =
-        sqlx::query_as("SELECT source, fetched_at FROM author_photos WHERE author_id = ?")
-            .bind(author_id)
-            .fetch_optional(pool)
-            .await?;
-    Ok(row.and_then(|(s, t)| AuthorPhotoSource::parse(&s).map(|src| (src, t))))
-}
-
-/// Upsert a photo row. Replaces any existing row (PRIMARY KEY conflict on
-/// `author_id`). `bytes` / `mime` are `None` for `'letter'` negative-cache
-/// markers; `url` is `None` for manual uploads.
-pub async fn upsert_author_photo(
-    pool: &SqlitePool,
-    author_id: i64,
-    source: AuthorPhotoSource,
-    url: Option<&str>,
-    mime: Option<&str>,
-    bytes: Option<&[u8]>,
-) -> Result<(), sqlx::Error> {
-    sqlx::query(
-        "INSERT INTO author_photos (author_id, source, url, mime, bytes, fetched_at)
-              VALUES (?, ?, ?, ?, ?, datetime('now'))
-         ON CONFLICT(author_id) DO UPDATE SET
-              source     = excluded.source,
-              url        = excluded.url,
-              mime       = excluded.mime,
-              bytes      = excluded.bytes,
-              fetched_at = excluded.fetched_at",
-    )
-    .bind(author_id)
-    .bind(source.as_str())
-    .bind(url)
-    .bind(mime)
-    .bind(bytes)
-    .execute(pool)
-    .await?;
-    Ok(())
-}
-
-/// Drop the cache row for an author so the next page view re-queues
-/// resolution. Used by admin DELETE.
-pub async fn delete_author_photo(pool: &SqlitePool, author_id: i64) -> Result<(), sqlx::Error> {
-    sqlx::query("DELETE FROM author_photos WHERE author_id = ?")
-        .bind(author_id)
-        .execute(pool)
-        .await?;
-    Ok(())
-}
-
-/// Delete an author by id and prevent reindex from re-creating the row.
-///
-/// F5.9-lite durable junk-author removal (issue #159). One transaction:
-/// look up the name (for the blocklist insert) and affected book ids
-/// (for the post-commit FTS refresh), drop the link rows, drop the
-/// `authors` row, then `INSERT OR IGNORE` the name into
-/// `ignored_authors` so the next `indexer::reindex` does not silently
-/// re-create the row via `resolve_or_insert_author`.
-///
-/// Returns the number of books that were un-linked (used by the admin
-/// confirmation modal to show "this affects N books"). Returns `Ok(0)`
-/// without touching the blocklist if `author_id` does not exist — a
-/// stale tab firing a second Delete must not leak a ghost row into
-/// `ignored_authors`.
-///
-/// FTS refresh runs best-effort *after* commit, mirroring
-/// [`upsert_metadata_overrides`]: a stale FTS row is fixed on the next
-/// reindex, but a refresh failure must not undo the admin's intent.
-pub async fn delete_author(pool: &SqlitePool, author_id: i64) -> Result<u64, sqlx::Error> {
-    let mut tx = pool.begin().await?;
-
-    let Some(name): Option<String> = sqlx::query_scalar("SELECT name FROM authors WHERE id = ?")
-        .bind(author_id)
-        .fetch_optional(&mut *tx)
-        .await?
-    else {
-        return Ok(0);
-    };
-
-    // Snapshot the affected book UUIDs *inside* the transaction by
-    // joining link → books. Doing the lookup as a uuid join (one bind
-    // parameter) instead of post-commit `WHERE id IN (?, ?, …)` keeps
-    // the post-commit FTS refresh from hitting SQLite's 999 bind-
-    // parameter cap on authors linked to >999 books — without that
-    // cap a giant-author delete would silently skip its FTS rebuild
-    // and leave the index stale.
-    let affected_uuids: Vec<String> = sqlx::query_scalar(
-        "SELECT b.uuid FROM books b
-         JOIN books_authors_link l ON l.book = b.id
-         WHERE l.author = ?",
-    )
-    .bind(author_id)
-    .fetch_all(&mut *tx)
-    .await?;
-
-    sqlx::query("DELETE FROM books_authors_link WHERE author = ?")
-        .bind(author_id)
-        .execute(&mut *tx)
-        .await?;
-
-    sqlx::query("DELETE FROM authors WHERE id = ?")
-        .bind(author_id)
-        .execute(&mut *tx)
-        .await?;
-
-    sqlx::query("INSERT OR IGNORE INTO ignored_authors(name) VALUES (?)")
-        .bind(&name)
-        .execute(&mut *tx)
-        .await?;
-
-    tx.commit().await?;
-
-    for uuid in &affected_uuids {
-        if let Err(e) = rebuild_fts_for_book(pool, uuid).await {
-            tracing::warn!(uuid = uuid.as_str(), error = %e, "books_fts rebuild after delete_author failed");
-        }
-    }
-
-    Ok(affected_uuids.len() as u64)
-}
 
 
 // -----------------------------------------------------------------------------
@@ -788,6 +594,10 @@ mod tests {
     };
     use crate::covers::delete_cover_files_for;
     use crate::helpers::stable_uuid;
+    use crate::author_photos_data::{
+        author_photo_status, delete_author, delete_author_photo, get_author_photo,
+        upsert_author_photo, AuthorPhotoSource,
+    };
     use crate::discovery::{get_author, get_series, get_tag_cloud, MAX_DISCOVERY_BOOKS};
     use crate::sync::{replace_books, sync_books, SyncPlan};
     use omnibus_shared::{Contributor, EbookMetadata, Identifier, MetadataOverrides};
